fix: Resolve circular imports
