+++ src/super_gradients/common/factories/__init__.py
@@ -8,7 +8,6 @@ from super_gradients.common.factories.optimizers_type_factory import OptimizersT
 from super_gradients.common.factories.samplers_factory import SamplersFactory
 from super_gradients.common.factories.transforms_factory import TransformsFactory
 from super_gradients.common.factories.activations_type_factory import ActivationsTypeFactory
-from super_gradients.common.factories.detection_modules_factory import DetectionModulesFactory
 
 
 __all__ = [
@@ -22,5 +21,4 @@ __all__ = [
     "ActivationsTypeFactory",
     "TypeFactory",
     "BaseFactory",
-    "DetectionModulesFactory",
 ]
